+++ packages/main/src/components/AnalyticalTable/hooks/useDragAndDrop.ts
@@ -5,7 +5,7 @@ const getColumnId = (column) => {
   return typeof column.accessor === 'string' ? column.accessor : column.id;
 };
 
-export const useDragAndDrop = (props, setColumnOrder, columnOrder, resizeInfo, columns: any[]) => {
+export const useDragAndDrop = (props, isRtl, setColumnOrder, columnOrder, resizeInfo, columns: any[]) => {
   const { onColumnsReordered } = props;
 
   const [dragOver, setDragOver] = useState('');
@@ -16,7 +16,7 @@ export const useDragAndDrop = (props, setColumnOrder, columnOrder, resizeInfo, c
         e.preventDefault();
         return;
       }
-      e.dataTransfer.setData('colId', e.currentTarget.dataset.columnId);
+      e.dataTransfer.setData('text', e.currentTarget.dataset.columnId);
     },
     [resizeInfo.isResizingColumn]
   );
@@ -34,7 +34,7 @@ export const useDragAndDrop = (props, setColumnOrder, columnOrder, resizeInfo, c
       setDragOver('');
 
       const droppedColId = e.currentTarget.dataset.columnId;
-      const draggedColId = e.dataTransfer.getData('colId');
+      const draggedColId = e.dataTransfer.getData('text');
       if (droppedColId === draggedColId) return;
 
       const internalColumnOrder = columnOrder.length > 0 ? columnOrder : columns.map((col) => getColumnId(col));
@@ -42,8 +42,7 @@ export const useDragAndDrop = (props, setColumnOrder, columnOrder, resizeInfo, c
       const draggedColIdx = internalColumnOrder.findIndex((col) => col === draggedColId);
 
       const tempCols = [...internalColumnOrder];
-
-      const targetIndex = droppedColIdx > draggedColIdx ? droppedColIdx - 1 : droppedColIdx;
+      const targetIndex = droppedColIdx > draggedColIdx ? (isRtl ? droppedColIdx : droppedColIdx - 1) : droppedColIdx;
 
       tempCols.splice(targetIndex, 0, tempCols.splice(draggedColIdx, 1)[0]);
       setColumnOrder(tempCols);