+++ packages/main/src/components/AnalyticalTable/hooks/useRowNavigationIndicator.tsx
@@ -11,9 +11,10 @@ const baseStyles = {
  */
 const Header = () => <div style={{ width: '6px' }} />;
 
-const Cell = ({ isNavigatedCell }) => {
+const Cell = (instance) => {
+  const { isNavigatedCell, webComponentsReactProperties } = instance;
   if (isNavigatedCell) {
-    return <div style={{ ...baseStyles, backgroundColor: ThemingParameters.sapList_SelectionBorderColor }} />;
+    return <div style={baseStyles} className={webComponentsReactProperties.classes.valueStateInformation} />;
   }
   return <div style={baseStyles} />;
 };
@@ -22,10 +23,7 @@ const Cell = ({ isNavigatedCell }) => {
  * TABLE HOOKS
  */
 const columnsDeps = (deps, { instance: { webComponentsReactProperties } }) => {
-  return [
-    ...deps,
-    webComponentsReactProperties.withNavigationHighlight
-  ];
+  return [...deps, webComponentsReactProperties.withNavigationHighlight];
 };
 const visibleColumnsDeps = (deps, { instance }) => [
   ...deps,