+++ packages/main/src/components/AnalyticalTable/hooks/useRowSelectionColumn.tsx
@@ -54,7 +54,7 @@ const Cell = ({ row, webComponentsReactProperties: { selectionBehavior, selectio
     return null;
   }
   if (selectionMode === TableSelectionMode.SINGLE_SELECT) {
-    return <div style={divStyle} onClick={handleCellClick} />;
+    return <div style={divStyle} onClick={handleCellClick} data-name="internal_selection_column" />;
   }
   return (
     <CheckBox
@@ -98,15 +98,16 @@ const columns = (currentColumns, { instance }) => {
   if (selectionMode === TableSelectionMode.NONE || selectionBehavior === TableSelectionBehavior.ROW_ONLY) {
     return currentColumns;
   }
+  const tableSelectionColumnWidth =
+    tableRef.current &&
+    parseInt(
+      getComputedStyle(tableRef.current).getPropertyValue(
+        `--${CssSizeVariablesNames.sapWcrAnalyticalTableSelectionColumnWidth}`
+      ),
+      10
+    );
+  const selectionColumnWidth = !isNaN(tableSelectionColumnWidth) ? tableSelectionColumnWidth : 47;
 
-  const selectionColumnWidth = tableRef.current
-    ? parseInt(
-        getComputedStyle(tableRef.current).getPropertyValue(
-          `--${CssSizeVariablesNames.sapWcrAnalyticalTableSelectionColumnWidth}`
-        ),
-        10
-      )
-    : 47;
   return [
     {
       id: '__ui5wcr__internal_selection_column',